SingleResultFuture moved to async
JAVA-1333
