+++ driver-async/src/main/com/mongodb/async/client/MappingIterable.java
@@ -21,7 +21,7 @@ import com.mongodb.Function;
 import com.mongodb.MongoException;
 import com.mongodb.async.MongoFuture;
 import com.mongodb.connection.SingleResultCallback;
-import com.mongodb.operation.SingleResultFuture;
+import com.mongodb.async.SingleResultFuture;
 
 import java.util.Collection;
