+++ driver-core/src/main/com/mongodb/protocol/CommandProtocol.java
@@ -25,7 +25,7 @@ import com.mongodb.connection.ResponseBuffers;
 import com.mongodb.diagnostics.Loggers;
 import com.mongodb.diagnostics.logging.Logger;
 import com.mongodb.operation.QueryFlag;
-import com.mongodb.operation.SingleResultFuture;
+import com.mongodb.async.SingleResultFuture;
 import com.mongodb.operation.SingleResultFutureCallback;
 import com.mongodb.protocol.message.CommandMessage;
 import com.mongodb.protocol.message.ReplyMessage;
