+++ driver-core/src/main/com/mongodb/protocol/GetMoreReceiveProtocol.java
@@ -19,7 +19,7 @@ package com.mongodb.protocol;
 import com.mongodb.async.MongoFuture;
 import com.mongodb.connection.Connection;
 import com.mongodb.connection.ResponseBuffers;
-import com.mongodb.operation.SingleResultFuture;
+import com.mongodb.async.SingleResultFuture;
 import com.mongodb.operation.SingleResultFutureCallback;
 import com.mongodb.protocol.message.ReplyMessage;
 import org.bson.codecs.Decoder;
