+++ driver-core/src/main/com/mongodb/protocol/QueryProtocol.java
@@ -26,7 +26,7 @@ import com.mongodb.connection.ServerDescription;
 import com.mongodb.diagnostics.Loggers;
 import com.mongodb.diagnostics.logging.Logger;
 import com.mongodb.operation.QueryFlag;
-import com.mongodb.operation.SingleResultFuture;
+import com.mongodb.async.SingleResultFuture;
 import com.mongodb.operation.SingleResultFutureCallback;
 import com.mongodb.protocol.message.QueryMessage;
 import com.mongodb.protocol.message.ReplyMessage;
