+++ driver-core/src/main/com/mongodb/protocol/ReplaceProtocol.java
@@ -25,7 +25,7 @@ import com.mongodb.connection.SingleResultCallback;
 import com.mongodb.diagnostics.Loggers;
 import com.mongodb.diagnostics.logging.Logger;
 import com.mongodb.operation.ReplaceRequest;
-import com.mongodb.operation.SingleResultFuture;
+import com.mongodb.async.SingleResultFuture;
 import com.mongodb.protocol.message.MessageSettings;
 import com.mongodb.protocol.message.ReplaceMessage;
 import com.mongodb.protocol.message.RequestMessage;
