+++ driver-core/src/main/com/mongodb/protocol/SendMessageCallback.java
@@ -20,7 +20,7 @@ import com.mongodb.MongoException;
 import com.mongodb.connection.Connection;
 import com.mongodb.connection.ResponseBuffers;
 import com.mongodb.connection.SingleResultCallback;
-import com.mongodb.operation.SingleResultFuture;
+import com.mongodb.async.SingleResultFuture;
 import org.bson.io.OutputBuffer;
 
 class SendMessageCallback<T> implements SingleResultCallback<Void> {
