+++ driver-core/src/main/com/mongodb/protocol/UnacknowledgedWriteResultCallback.java
@@ -21,7 +21,7 @@ import com.mongodb.MongoNamespace;
 import com.mongodb.async.MongoFuture;
 import com.mongodb.connection.Connection;
 import com.mongodb.connection.SingleResultCallback;
-import com.mongodb.operation.SingleResultFuture;
+import com.mongodb.async.SingleResultFuture;
 import com.mongodb.operation.SingleResultFutureCallback;
 import com.mongodb.protocol.message.RequestMessage;
 import org.bson.io.OutputBuffer;
