+++ driver-core/src/main/com/mongodb/protocol/UpdateCommandProtocol.java
@@ -23,7 +23,7 @@ import com.mongodb.async.MongoFuture;
 import com.mongodb.connection.Connection;
 import com.mongodb.connection.SingleResultCallback;
 import com.mongodb.diagnostics.Loggers;
-import com.mongodb.operation.SingleResultFuture;
+import com.mongodb.async.SingleResultFuture;
 import com.mongodb.operation.UpdateRequest;
 import com.mongodb.operation.WriteRequest;
 import com.mongodb.protocol.message.MessageSettings;
