+++ driver-core/src/main/com/mongodb/protocol/WriteCommandProtocol.java
@@ -24,7 +24,7 @@ import com.mongodb.connection.ByteBufferOutputBuffer;
 import com.mongodb.connection.Connection;
 import com.mongodb.connection.ResponseBuffers;
 import com.mongodb.connection.SingleResultCallback;
-import com.mongodb.operation.SingleResultFuture;
+import com.mongodb.async.SingleResultFuture;
 import com.mongodb.operation.SingleResultFutureCallback;
 import com.mongodb.operation.WriteRequest;
 import com.mongodb.protocol.message.BaseWriteCommandMessage;
